requirements.txt
amqp==5.3.1
billiard==4.2.1
blinker==1.9.0
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.4
MarkupSafe==3.0.2
mcrcon==0.7.0
packaging==25.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==6.2.0
requests==2.32.4
six==1.17.0
SQLAlchemy==2.0.41
stripe==12.4.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
py-cord==2.5.0
mcstatus==11.0.1
PyJWT==2.7.0
bcrypt==4.0.1